LocalAI/requirements.txt
chromadb==0.5.5
colorama==0.4.6
gTTS==2.5.3
ollama==0.3.3
openai==1.53.0
playsound==1.3.0
psycopg==3.2.1
python-dotenv==1.0.1
python-socketio==5.11.4
SpeechRecognition==3.10.4
tqdm==4.66.5
